requirements.txt
snowflake-connector-python[pandas]==3.15.0
duckdb>=1.1.1,<2.0.0
streamlit==1.40.0
pandas==2.2.3
python-dotenv==1.0.1
plotly==5.17.0